+++ test/random_unicode_range_test.dart
@@ -9,17 +9,22 @@ import 'package:test/test.dart';
 void main() {
   group('RandomUnicodeRange -', () {
     test('constructor - excessive', () {
-      final r = RandomUnicodeRange(-3, 0x77777777);
+      final r = RandomUnicodeRange(min: -3, max: 0x77777777);
       expect(r.min == 0 && r.max == 0x10FFFF, true);
     });
     test('constructor - trivial', () {
-      final r = RandomUnicodeRange(0x21);
+      final r = RandomUnicodeRange(min: 0x21);
       expect(r.min, 0x21);
       expect(r.max, 0x21);
     });
-    test('nextCharCode', () {
-      final r = RandomUnicodeRange(0x10, 0x20);
-      final c = r.nextCharCode(RandomUnicode.defaultRandom.nextInt);
+    test('nextCharCode - list', () {
+      final r = RandomUnicodeRange(list: [0x10, 0x20, 0x30]);
+      final c = r.nextCharCode();
+      expect(c == 0x10 || c == 0x20 || c == 0x30, true);
+    });
+    test('nextCharCode - min/max', () {
+      final r = RandomUnicodeRange(min: 0x10, max: 0x20);
+      final c = r.nextCharCode();
       expect(c >= r.min && c <= r.max, true);
     });
   });