+++ CHANGELOG.md
@@ -1,6 +1,15 @@
+## 0.2.0
+
+- Breaking change: passing just Random object to RandomUnicode constructors
+- Bug fix: added Random object as the parameter to non-default constructors
+- Breaking change: passing Random object to RandomUnicodeRange constructor
+- RandomUnicodeRange can hold a non-contiguous list of values
+- Bug fix: when calling random.nextInt(max), assumed that max was included 
+
 ## 0.1.1
 
 - Fixed README and changed the homepage
+
 ## 0.1.0
 
 - Initial version.